api/requirements.txt
uvicorn==0.22.0
fastapi==0.96.0
passlib
python-jose==3.3.0
bcrypt
